+++ include/RtFrameAccumulator.h
@@ -27,17 +27,18 @@ public:
 	// Adds one renderPass() result (width*height radiance samples) into the
 	// running sum. sampleRadiance must match the current width()/height().
 	// primaryHitMask, if provided (see CpuPathTracer::renderPass), must also
-	// match size - per-pixel hit counts are tracked alongside the radiance sum
-	// so callers can tell "this pixel's primary ray has never hit geometry"
-	// (pure background) apart from "this pixel is on a silhouette edge and
-	// flips between hit/miss across AA-jittered samples). primaryAlbedo/
+	// match size - per-pixel primary coverage is tracked alongside the
+	// radiance sum so callers can tell "this pixel's primary ray has never hit
+	// geometry" (pure background) apart from "this pixel is on a silhouette
+	// edge or shadow catcher and only partially covers the raster backdrop
+	// across AA-jittered samples). primaryAlbedo/
 	// primaryNormal, if provided, are accumulated the same way (running sum,
 	// divided by sampleCount at resolve time) - see
 	// CpuPathTracer::renderPass()'s outPrimaryAlbedo/outPrimaryNormal for
 	// what these represent (OIDN denoiser guide buffers, RtDenoiser::
 	// denoise()).
 	void accumulate(const std::vector<glm::vec3>& sampleRadiance,
-		const std::vector<uint8_t>* primaryHitMask = nullptr,
+		const std::vector<float>* primaryHitMask = nullptr,
 		const std::vector<glm::vec3>* primaryAlbedo = nullptr,
 		const std::vector<glm::vec3>* primaryNormal = nullptr);
 
@@ -51,10 +52,9 @@ public:
 	std::vector<glm::vec3> resolveAlbedo() const;
 	std::vector<glm::vec3> resolveNormal() const;
 
-	// Per-pixel count of accumulated samples whose primary ray hit geometry
-	// (see accumulate()'s primaryHitMask). Empty if accumulate() was never
-	// called with a mask.
-	const std::vector<uint32_t>& hitCounts() const { return _hitCounts; }
+	// Per-pixel accumulated primary coverage/alpha (see accumulate()'s
+	// primaryHitMask). Empty if accumulate() was never called with a mask.
+	const std::vector<float>& hitCounts() const { return _hitCounts; }
 
 	int width() const { return _width; }
 	int height() const { return _height; }
@@ -66,6 +66,6 @@ private:
 	std::vector<glm::vec3> _sum;
 	std::vector<glm::vec3> _albedoSum;
 	std::vector<glm::vec3> _normalSum;
-	std::vector<uint32_t> _hitCounts;
+	std::vector<float> _hitCounts;
 	uint32_t _sampleCount = 0;
 };
